refactor(types)!: ID ranges and value structure
change convention for managed/global and private/limited types
reduce scope range for meta and regular managed types
use address-like identifier for private types

+++ mptcore/convert/value_convert.c
@@ -47,11 +47,11 @@ extern int mpt_value_convert(const MPT_STRUCT(value) *val, int type, void *dest)
 	/* try specialized converter for type */
 	if ((conv = mpt_data_converter(val->_type))) {
 		if ((ret = conv(src, type, dest)) >= 0) {
-			return val->_type == type ? 0 : 3;
+			return val->_type == (unsigned int) type ? 0 : 3;
 		}
 	}
 	/* exact primitive type match */
-	if (val->_type == type) {
+	if (val->_type == (unsigned int) type) {
 		const MPT_STRUCT(type_traits) *traits;
 		if (!(traits = mpt_type_traits(type))) {
 			return MPT_ERROR(BadArgument);
